requirements.txt
sklearn
numpy
PyYAML
configargparse
smplx
pillow
opencv-python
tqdm
pyrender
trimesh
scipy
open3d
times
DateTime
tensorboard
pandas
ipdb
chumpy
torchgeometry
openmesh
mesh-to-sdf
